Set rwb.app to the instance of the application object
This is an experiment. The idea is, I don't want to have to pass a
reference to the app everywhere that needs it. Now each module can
import rwb and use rwb.app to get a handle to the app (so they can
get theme colors and fonts, settings, etc).